add error responce

+++ DreamNekosConnect.Lib/Services/InterestService.cs
@@ -1,5 +1,5 @@
-﻿using DreamNekosConnect.Lib.Entities;
-using DreamNekosConnect.Lib.Providers;
+﻿using DreamNekos.API.Helpers;
+using DreamNekosConnect.Lib.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace DreamNekosConnect.Lib.Services
@@ -17,6 +17,7 @@ namespace DreamNekosConnect.Lib.Services
         public InterestEntity CreateInterest(string name, Guid? interestTypeId)
         {
             var interestType = _dbContext.InterestType.FirstOrDefault(x => x.Id == interestTypeId);
+            if (interestType == null) throw new ElementNotFoundException("A interest type with such an ID does not exist.");
             InterestEntity newInterest = new InterestEntity{ Name = name, InterestType = interestType }; 
             _dbContext.Add(newInterest);
             _dbContext.SaveChanges();
@@ -32,18 +33,23 @@ namespace DreamNekosConnect.Lib.Services
             var interest = _dbContext.Interests
                 .Include(x => x.InterestType)
                 .First(x => x.Id == id);
+            if (interest == null) throw new ElementNotFoundException("A interest with such an ID does not exist.");
             interest.Name = name ?? interest.Name;
 
             if(interestTypeId != null)
             {
                 var interestType = _dbContext.InterestType.FirstOrDefault(x => x.Id == interestTypeId);
+                if (interestType == null) throw new ElementNotFoundException("A interest type with such an ID does not exist.");
                 interest.InterestType = interestType;
             }
 
             _dbContext.SaveChanges();
             return interest;
         }
-        public void DeleteInterest(InterestEntity interest) {
+        public void DeleteInterest(Guid interestId) {
+            var interest = _dbContext.Interests
+               .First(x => x.Id == interestId);
+            if (interest == null) throw new ElementNotFoundException("A interest with such an ID does not exist.");
             _dbContext.Interests.Remove(interest);
             _dbContext.SaveChanges();
         }
